node: move gRPC test_util in grpc folder
