+++ silkworm/node/execution/grpc/client/endpoint/insertion_test.cpp
@@ -25,7 +25,7 @@
 #include <silkworm/node/test_util/fixture.hpp>
 #include <silkworm/node/test_util/sample_blocks.hpp>
 
-#include "../../../test_util/sample_protos.hpp"
+#include "../../test_util/sample_protos.hpp"
 
 namespace silkworm::execution::grpc::client {
