+++ silkworm/node/execution/grpc/common/block_test.cpp
@@ -24,7 +24,7 @@
 #include <silkworm/node/test_util/fixture.hpp>
 #include <silkworm/node/test_util/sample_blocks.hpp>
 
-#include "../../test_util/sample_protos.hpp"
+#include "../test_util/sample_protos.hpp"
 
 namespace silkworm::execution::grpc {
