+++ silkworm/node/execution/grpc/server/endpoint/getters_test.cpp
@@ -26,7 +26,7 @@
 #include <silkworm/node/test_util/fixture.hpp>
 #include <silkworm/node/test_util/sample_blocks.hpp>
 
-#include "../../../test_util/sample_protos.hpp"
+#include "../../test_util/sample_protos.hpp"
 
 namespace silkworm::execution::grpc::server {
